Fix unstable test
See https://rt.cpan.org/Public/Bug/Display.html?id=109261

This test depended on the order of elements in a hash. Since Perl 5.17.6, this
order is randomized, resulting in the test sometimes succeeding and sometimes
failing in recent Perl versions.

This is solved by adjusting the test so that there is only one violation of the
contract.